+++ src/seedcase_sprout/core/create_resource_properties.py
@@ -6,11 +6,12 @@ from seedcase_sprout.core.check_properties import (
 from seedcase_sprout.core.create_relative_resource_data_path import (
     create_relative_resource_data_path,
 )
+from seedcase_sprout.core.paths import PackagePath
 from seedcase_sprout.core.properties import ResourceProperties
 
 
 def create_resource_properties(
-    path: Path, properties: ResourceProperties
+    properties: ResourceProperties, path: Path | None = None
 ) -> ResourceProperties:
     """Creates a valid properties object for the specified resource.
 
@@ -19,12 +20,12 @@ def create_resource_properties(
     them to be added to the `datapackage.json` file.
 
     Args:
-        path: The path to the resource `id` folder; use `PackagePath().resource()`
-            to provide the correct path.
         properties: The properties of the resource; must be given as a
             `ResourceProperties` object following the Data Package specification.
             See the `ResourceProperties` help documentation for details
             on what can or needs to be filled in.
+        path: The path to the resource `id` folder; use `PackagePath().resource()`
+            to provide the correct path.
 
     Returns:
         The properties object, verified and updated.
@@ -54,15 +55,16 @@ def create_resource_properties(
             # TODO: Update after converting to "local-first"
             # Create resource properties
             # sp.create_resource_properties(
-            #     path=temp_path / "1" / "resources" / "1",
             #     properties=sp.ResourceProperties(
             #         name="new-resource-name",
             #         path="data.parquet",
             #         title="Resource Title",
             #         description="This resource contains data about...",
             #     ),
+            #     path=temp_path / "1" / "resources" / "1",
             # )
         ```
     """
+    path = path or PackagePath().resource(properties.name)
     properties.path = str(create_relative_resource_data_path(path))
     return check_resource_properties(properties)